Refactor trainer and evaluators to work with CLU DatasetIterators rather than tf.data.Dataset objects.
This doesn't change the input pipeline, it just changes the type of object that it returns. CLU DatasetIterators are checkpointable, although if the iterator is a TfDatasetIterator (as it is currently), checkpointing becomes prohibitively expensive.

The goal is to switch to a pipeline that is 1) fully deterministic and 2) preemption-proof. As long as the output of the pipeline is a DatasetIterator, we don't need to update anything else from the code.

PiperOrigin-RevId: 463551372