TreeSummarizer class used to summarize arrays in the gradients tree/optimizer state in arbitrary ways.
Before, we could plot the norm of such arrays, but in some cases we want to be able to plot other types of summary values, or even multiple values per array (e.g. if we want to individually plot the norm of the params of each expert).

See the tests for some examples.

PiperOrigin-RevId: 506610280