Add options to control rounding in compute_capacity.
When the number of tokens is small, or the number of experts is big, ceiling to the next divisor of 4 can make that the actual capacity factor is much larger than the one indicated. This can lead to wrong conclusions when comparing different capacity factors.

Example: Take a model with patch size 32, trained on a resolution of 224 pixels. We take 4 images per group, that makes a group_size of 4 * (224/32)^2 = 196.

With a capacity_factor = 2.0, the capacity per expert with 32 experts should be 12.25. After ceiling to the next multiple of 4 we get a capacity per expert of 16.
This results in an actual capacity factor of 16*32/196 = 2.61, which is almost 31% larger than the target.

PiperOrigin-RevId: 514754784