Fix save_checkpoint bug for partitioned models.
The save_checkpoint code assumed that all processes created the same Index. The index contains which checkpoint shards contain each of the array's slices.

The problem is that the function _create_map_slicend_to_array_shards may return dictionaries with the same keys, but different orders in each process. This is because the global_shards attribute of jax.Array can return the shards in different orders in each process.

The result is that two processes might decide to write the same slice of an array to one of their corresponding checkpoint shards, resulting in a corruption of the train state.

PiperOrigin-RevId: 514881151