Report training flops and optimal device time computed by XLA while training.
This is useful to plot loss/accuracy w.r.t. total training flops or total device time.

Be aware that if your model uses XLA while_loop, these costs might be inaccurate, since only the cost of a single step is used. But this is not the case for most of our models.

PiperOrigin-RevId: 618232607